ssh-key: refactor module structure
Refactors the algorithm-specific modules from the `algorithm` module to
per-algorithm modules under `public` and `private`.
